Move file selector to the top of the screen
ToolBelt is split into an upper and a lower part where the upper one
consists of the file selector and a settings button. The lower toolbelt
has the tile, mode and layer selectors (the latter has become wider) and
the mouse position display. This change gives more space to file name
and layer name labels.